refactor: nextstep 패키지로 기존 domain 클래스들 이동
