+++ src/main/java/ladder/domain/engine/LadderResults.java
@@ -1,6 +1,6 @@
 package ladder.domain.engine;
 
-import ladder.domain.Player;
+import ladder.domain.nextstep.Player;
 
 import java.util.Collections;
 import java.util.HashMap;
