+++ src/main/java/ladder/domain/nextstep/NextStepLineGenerator.java
@@ -1,0 +1,4 @@
+package ladder.domain.nextstep;
+
+public class NextStepLineGenerator {
+}
