+++ src/main/java/ladder/domain/nextstep/RandomGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.nextstep;
 
 import java.util.Random;
