+++ src/test/java/ladder/NextStepLadderTest.java
@@ -1,14 +1,14 @@
 package ladder;
 
-import ladder.domain.Ladder;
+import ladder.domain.nextstep.NextStepLadder;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-public class LadderTest {
+public class NextStepLadderTest {
     @Test
     public void Ladder_높이_Test(){
-        assertThatThrownBy(() -> new Ladder(3,0))
+        assertThatThrownBy(() -> new NextStepLadder(3,0))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("사다리 높이는 1이상이어야 합니다.");
     }
