+++ src/test/java/ladder/NextStepLineTest.java
@@ -1,14 +1,14 @@
 package ladder;
 
-import ladder.domain.Line;
+import ladder.domain.nextstep.NextStepLine;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class LineTest {
+public class NextStepLineTest {
     @Test
     public void Line_Point_수_Test(){
-        Line line = new Line(10);
+        NextStepLine line = new NextStepLine(10);
         assertThat(line.points().size()).isEqualTo(10);
     }
 }
