+++ src/test/java/ladder/PlayerTest.java
@@ -1,6 +1,6 @@
 package ladder;
 
-import ladder.domain.Player;
+import ladder.domain.nextstep.Player;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
