+++ src/test/java/ladder/PointTest.java
@@ -1,6 +1,6 @@
 package ladder;
 
-import ladder.domain.Point;
+import ladder.domain.nextstep.Point;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
